Misc cleanup -asm files no longer need _ASM in the name if they share a name with a cpp file. -Fixed a few warnings -Build and Run now tracks errors correctly.

+++ kernel/arch/i386/kernel/Terminal.cpp
@@ -69,7 +69,7 @@ namespace kernel
         else if(c == '\r')
         {
             m_Column = 0;
-            for(int i = 0; i < s_VGAWidth; ++i)
+            for (size_t i = 0; i < s_VGAWidth; ++i)
             {
                 PutChar(' ');
             }
